User Management #105

+++ conf/evolutions/default/3.sql
@@ -2,12 +2,12 @@
 
 # --- !Ups
 CREATE TABLE event (
-  id                 	integer NOT NULL PRIMARY KEY,
+  id                 	BIGINT NOT NULL PRIMARY KEY,
   type					varchar(255) NOT NULL,
-  thinking_session		integer NOT NULL REFERENCES thinking_session(id),
-  hat                  	integer NOT NULL REFERENCES hat(id),
-  user					integer REFERENCES user(id),
-  card					integer REFERENCES card(id),
+  thinking_session		BIGINT NOT NULL REFERENCES thinking_session(id),
+  hat                  	BIGINT NOT NULL REFERENCES hat(id),
+  user					BIGINT REFERENCES user(id),
+  card					BIGINT REFERENCES card(id),
   time					datetime DEFAULT NOW()
 );
